+++ packages/ic-management/src/index.ts
@@ -1,3 +1,8 @@
+export type {
+  canister_log_record,
+  fetch_canister_logs_result,
+  snapshot_id,
+} from "../candid/ic-management";
 export { ICManagementCanister } from "./ic-management.canister";
 export * from "./types/canister.options";
 export * from "./types/ic-management.params";