+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Added `memoToNeuronSubaccount` and `memoToNeuronAccountIdentifier`.
 - Support new neuron field `voting_power_refreshed_timestamp_seconds`.
 - Add support for fetching the canister logs in `@dfinity@ic-management`.
+- Add support for snapshot features in `@dfinity@ic-management`.
 - Add `nowInBigIntNanoSeconds` to `@dfinity/utils`, a trivial function that is actually used across all our dapps.
 
 ## Build